% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cwr_tidy}
\alias{cwr_tidy}
\title{CWR tidy}
\format{
A data frame with 593 rows and 19 variables:
\describe{
\item{year}{Year of encounter}
\item{date}{Date of encounter}
\item{encounter_sequence}{Encounter sequence}
\item{encounter_number}{Encounter number}
\item{begin_time}{Encounter start time}
\item{end_time}{Encounter end time}
\item{duration}{Duration of encounter}
\item{vessel}{Name of vessel(s) used for encounter observation}
\item{observers}{Names of observers}
\item{pods_or_ecotype}{Pod(s) and/or ecotype observed}
\item{ids_encountered}{IDs of whales encountered}
\item{location}{Description of location}
\item{begin_latitude}{Latitude encounter began}
\item{end_latitude}{Latitude encounter ended}
\item{begin_longitude}{Longitude encounter began}
\item{end_longitude}{Longitude encounter ended}
\item{encounter_summary}{Summary of encounter}
\item{nmfs_permit}{Permits for photos}
\item{link}{URL to encounter webpage}
...
}
}
\source{
\url{https://www.whaleresearch.com/}
}
\usage{
cwr_tidy
}
\description{
Tidy Center for Whale Research data from 2017 to 2023.
}
\keyword{datasets}
