% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cwr_untidy}
\alias{cwr_untidy}
\title{CWR untidy}
\format{
A data frame with 603 rows and 47 variables that need to be
wrangled to reduce the redundancy in column names.
}
\source{
\url{https://www.whaleresearch.com/}
}
\usage{
cwr_untidy
}
\description{
Untidy Center for Whale Research data from 2017 to 2023.
}
\keyword{datasets}
