% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_cwr.R
\name{get_all_encounters}
\alias{get_all_encounters}
\title{Make a Dataframe of Multiple CWR Encounters From One Year}
\usage{
get_all_encounters(years)
}
\arguments{
\item{years}{Numeric years from 2017 to present day.}
}
\value{
Dataframe with one row per encounter.
}
\description{
Getting all encounters from 2017 to 2023 makes about 30 minutes.
The resulting dataframe will need to be wrangled and cleaned. See
\verb{tidy_encounters().}
}
