% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_cwr.R
\name{get_encounter_data}
\alias{get_encounter_data}
\title{Extract CWR Data from Encounter Webpage}
\usage{
get_encounter_data(encounter_url)
}
\arguments{
\item{encounter_url}{Character url for CWR encounter.}
}
\value{
Character vector to be input to \code{parse_encounter()}.
}
\description{
Extract CWR Data from Encounter Webpage
}
\examples{
get_encounter_data("https://www.whaleresearch.com/2023-13")
}
