% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{href}
\alias{href}
\title{Wrap url with html \if{html}{\out{<a>}} tag}
\usage{
href(url, text = url)
}
\arguments{
\item{url}{URL of the page the link goes to.}

\item{text}{Character string to display. Defaults to url.}
}
\value{
Character string with \if{html}{\out{<a>}} tag.
}
\description{
Creates an html hyperlink with the \if{html}{\out{<a>}}
tag and target attribute set to open the link in a new tab.
}
