% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R
\name{make_dt}
\alias{make_dt}
\title{Make a DT}
\usage{
make_dt(data)
}
\arguments{
\item{data}{CWR tidy dataframe.}
}
\value{
A datatable object.
}
\description{
Creates an interactive DataTable using the R interface to the
JavaScript library DataTables
}
\examples{
make_dt(orcas::cwr_tidy[1:5, ])
}
