% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_cwr.R
\name{make_encounter_df}
\alias{make_encounter_df}
\title{Make a Dataframe of Multiple CWR Encounters}
\usage{
make_encounter_df(years, max_urls = Inf)
}
\arguments{
\item{years}{Numeric years from 2017 to present day.}

\item{max_urls}{Number of urls to extract (from newest to oldest
for each year). Defaults to \code{Inf}.}
}
\value{
Dataframe with one row per encounter. This dataframe will
need to be tidied. See data-raw/data_cwr.R for the data cleaning
script used to clean 2017:2024 data.
}
\description{
Getting all encounters from 2017 to 2024 takes about 40 minutes.
The resulting dataframe will need to be wrangled and cleaned. See
how it was tidied in \code{data-raw/data_cwr.R}
}
\examples{
\dontrun{
make_encounter_df(years = 2022:2024, max_urls = 1)
}
}
