% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{make_leaflet}
\alias{make_leaflet}
\title{Make leaflet map of CWR encounters with Stamen.Terrain tiles and
whale icons}
\usage{
make_leaflet(data)
}
\arguments{
\item{data}{Tidy dataframe with latitude and longitude columns}
}
\value{
Leaflet map
}
\description{
Make leaflet map of CWR encounters with Stamen.Terrain tiles and
whale icons
}
\examples{
{
df <- data.frame(
  encounter_number = c(14, 13),
  date = c("2023-04-08", "2023-03-25"),
  duration = c("4980s (~1.38 hours)", "23700s (~6.58 hours)"),
  pod_ecotype = c("J, L", "J"),
  location = c("East Sooke", "Haro Strait"),
  begin_latitude = c(48.3108329772949, 48.5881652832031),
  begin_longitude = c(-123.664001464844, -123.203163146973),
  end_latitude = c(48.300666809082, 48.7113342285156),
  end_longitude = c(-123.747329711914, -123.272163391113),
  encounter_summary = c(
    "While out scanning the western...",
    "After receiving reports of J pod ..."
  ),
  link = c(
    "https://www.whaleresearch.com/2023-14",
    "https://www.whaleresearch.com/2023-13"
  )
)
  make_leaflet(df)
}
}
