% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_cwr.R
\name{parse_encounter}
\alias{parse_encounter}
\title{Parse CWR Encounter Data}
\usage{
parse_encounter(encounter_data)
}
\arguments{
\item{encounter_data}{Character vector from \code{get_encounter_data()}.}
}
\value{
Dataframe with one row of encounter data parsed into
columns.
}
\description{
Parse CWR Encounter Data
}
\examples{
get_encounter_data("https://www.whaleresearch.com/2022-13") |>
  parse_encounter()
}
