% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_cwr.R
\name{save_encounter_df}
\alias{save_encounter_df}
\title{Save encounter dataframe to .Rdata}
\usage{
save_encounter_df(df)
}
\arguments{
\item{df}{Dataframe to save to data folder.}
}
\description{
Save encounter dataframe to .Rdata
}
