% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_cwr.R
\name{get_encounter_links}
\alias{get_encounter_links}
\title{Extract Link for Each Encounter}
\usage{
get_encounter_links(
  landing_url = "https://www.whaleresearch.com/2022encounters",
  website_type = "current",
  max_urls = Inf
)
}
\arguments{
\item{landing_url}{Character url to CWR encounter landing page.}

\item{website_type}{Either "current" or "archive" depending on whether the
encounter landing page is on the CWR current website or archive website.
Defaults to "current".}

\item{max_urls}{Number of urls to extract. Defaults to \code{Inf}.}
}
\value{
Character vector of urls
}
\description{
Extract encounter links from
\href{https://www.whaleresearch.com/2022encounters}{Center for Whale
Research (CWR)} encounter landing pages.

2022 and 2023 encounters are on their current website while 2017 - 2021
encounters are on their
\href{https://whaleresearch.wixsite.com/archives}{archival website}.
}
\examples{
library(rvest)
get_encounter_links("https://www.whaleresearch.com/2022encounters",
                    max_urls = 10)
}
