% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_cwr.R
\name{make_encounter_df}
\alias{make_encounter_df}
\title{Make a Dataframe of Multiple CWR Encounters From One Year}
\usage{
make_encounter_df(encounter_urls)
}
\arguments{
\item{encounter_urls}{Character vector of CWR urls. Pipe
\code{get_encounter_links} to this function.}
}
\value{
Dataframe with one row per encounter.
}
\description{
Make a Dataframe of Multiple CWR Encounters From One Year
}
\examples{
# Piping the `get_encounter_links()` function
get_encounter_links(year = 2023, max_urls = 2) |>
  make_encounter_df()
}
