% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{make_leaflet}
\alias{make_leaflet}
\title{Make leaflet map of CWR encounters with Stamen.Terrain tiles and
whale icons}
\usage{
make_leaflet(data)
}
\arguments{
\item{data}{Tidy dataframe with latitude and longitude columns}
}
\value{
Leaflet map
}
\description{
Make leaflet map of CWR encounters with Stamen.Terrain tiles and
whale icons
}
\examples{
  make_leaflet(orcas::cwr_tidy[1:5, ])
}
